requirements.txt
flask==1.0.2
numpy==1.14.5
pandas
pycrypto==2.6.1
